+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Changes to main, not yet committed to stable
+- Avatar update issue: other users can also see the update when an Avatar is updated by the user.
+
 
 *Please add a brief description of any changes to be tested and any migrations to be performed here.*